PgpBlockViewAttachmentsModule
